Add test infra and initial pinger test
Check out cockpit.git to get access to its test VM images and Python API
for using it. Use that to build a test VM with the distro's cockpit
package and our locally built cockpit-examples RPM.

As a first test, just check that the package is correctly installed.
(No browser test yet)